services/cancer-agent/requirements.txt
fastapi[standard]
pydantic
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
pandas
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow
python-dotenv
